Update trelica_2D
Due to the changes in the activities, I removed the time display, because, in this algorithm, there will be no iterations to vary the properties of the elements or the external forces anymore. This algorithm will be adapted to optimization, so the time display will happen in the BSA file, where the iterations will occur.