Updated files and added generate_truss_data
Added generate_truss_data.py which generates an excel file (e.g., output_100000.xlsx) and a sqlite3 database ("output_100000.db") with random data for the 2D 7-bar truss. This data will be used to create a surrogate model with machine learning and then the surrogate will be used to optimize the structure, instead of using the original model (which for large structures can be computationally costly).

A few files were minorly updated.